+++ CarpoolApp/CarpoolApp/Models/UsputniGradovi.cs
@@ -8,7 +8,7 @@ namespace CarpoolApp.Models
     public class UsputniGradovi
     {
         public int UsputniGradoviID { get; set; }
-        public double CijenaUsputni { get; set; }
+        public decimal CijenaUsputni { get; set; }
         public int VoznjaID { get; set; }
         public Voznja Voznja { get; set; }
         public int GradID { get; set; }
